+++ features/openVault/components/OpenVaultForm.tsx
@@ -6,12 +6,12 @@ import { pick } from 'helpers/pick'
 import { useSelectFromContext } from 'helpers/useSelectFromContext'
 import React from 'react'
 
-import { Allowance } from './formSections/Allowance'
-import { AllowanceStatus } from './formSections/AllowanceStatus'
-import { OpenVaultErrors } from './formSections/OpenVaultErrors'
-import { OpenVaultTitle } from './formSections/OpenVaultTitle'
-import { OpenVaultWarnings } from './formSections/OpenVaultWarnings'
-import { StatusCard } from './formSections/Status'
+import { Allowance } from './form/Allowance'
+import { AllowanceStatus } from './form/AllowanceStatus'
+import { OpenVaultErrors } from './form/OpenVaultErrors'
+import { OpenVaultTitle } from './form/OpenVaultTitle'
+import { OpenVaultWarnings } from './form/OpenVaultWarnings'
+import { StatusCard } from './form/Status'
 import { OpenVaultButton } from './OpenVaultButton'
 import { OpenVaultConfirmation, OpenVaultStatus } from './OpenVaultConfirmation'
 import { OpenVaultEditing } from './OpenVaultEditing'
